Add more serializer tests
